Extract spotlights dropdown to separate component
